made changes according to the PR
