+++ auth-connect-frontend/package.json
@@ -27,6 +27,7 @@
     "@types/redux": "^3.6.0",
     "@types/redux-thunk": "^2.1.0",
     "@types/styled-components": "^5.1.7",
+    "babel-plugin-styled-components": "^1.12.0",
     "typescript": "^4.1.5"
   }
 }
