+++ auth-connect-frontend/pages/index.tsx
@@ -1,4 +1,4 @@
-import Head from "next/head";
+import { NextPage } from "next";
 import styled from "styled-components";
 import { COLORS } from "../src/colors/constants";
 import Layout from "../src/components/shared/layout";
@@ -6,22 +6,20 @@ import Layout from "../src/components/shared/layout";
 const HeadingWrapper = styled.div`
   display: flex;
   justify-content: center;
+  align-items: center;
+  min-height: 80vh;
   color: ${COLORS.BLACK};
   font-size: 30px;
   font-weight: 600;
 `;
-const Home = () => {
+
+const Index: NextPage = () => {
   return (
     <>
-      <Head>
-        <title>Connect Authentication SDK</title>
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <Layout title={"Home"} />
-
+      <Layout title="Home" />
       <HeadingWrapper>Connect Authentication SDK</HeadingWrapper>
     </>
   );
 };
 
-export default Home;
+export default Index;
