+++ auth-connect-frontend/src/colors/constants.ts
@@ -1,6 +1,6 @@
 export const COLORS = {
   BACKGROUND: "#FCFCFC",
-  BLACK: "#333333",
+  BLACK: "#000000",
   BORDER: "#E2EBF8",
   BOX_SHADOW: "rgba(226, 235, 248, 0.15)",
   BOX_SHADOW_HARD: "rgba(226, 235, 248, 0.5)",
